fix(credentials): correct v3 keyblock width in CCache.Marshal
The version-3 keyblock encodes the key type as two 16-bit integers
(matching parseCredential), not a 16-bit value followed by a 32-bit one.
Without this, Marshal was not a true inverse of Unmarshal for v3 caches.

Add TestMarshalRoundTripV3, which builds a v3 cache and asserts its bytes
survive an Unmarshal/Marshal round trip with the key type intact.

+++ credentials/ccache_marshal.go
@@ -200,14 +200,16 @@ func (c *CCache) writePrincipalTo(b *bytes.Buffer, p principal, endian *binary.B
 }
 
 // writeKeyblock serializes a credential's encryption key: a 16-bit key type
-// (repeated as a 32-bit value in version 3), then a length-prefixed key value.
+// (repeated as a second 16-bit value in version 3), then a length-prefixed key
+// value. This mirrors parseCredential, which reads the key type as two 16-bit
+// integers for version 3.
 func (c *CCache) writeKeyblock(b *bytes.Buffer, key types.EncryptionKey, endian *binary.ByteOrder) error {
 	if err := binary.Write(b, *endian, uint16(key.KeyType)); err != nil {
 		return err
 	}
 
 	if c.Version == 3 {
-		if err := binary.Write(b, *endian, uint32(key.KeyType)); err != nil {
+		if err := binary.Write(b, *endian, uint16(key.KeyType)); err != nil {
 			return err
 		}
 	}
